Syntax error fixed,new color and side bar size for campaign,ui improved. Campaign animation doesnt work
