+++ src/components/Campaign.css
@@ -1,10 +1,23 @@
 :root {
-  --primary-color: #333;
+  --primary-color: #000; 
   --background-color: #fff;
   --border-color: #e1e1e1;
   --highlight-color: #007bff;
   --input-background: #f5f5f5;
   --input-border-radius: 4px;
+  --sidebar-width: 220px; 
+  --sidebar-font-size: 15px; 
+}
+
+@keyframes slideInFromLeft {
+  0% {
+    transform: translateX(-100%); 
+    opacity: 0; 
+  }
+  100% {
+    transform: translateX(0); 
+    opacity: 1; 
+  }
 }
 
 body, html {
@@ -16,15 +29,16 @@ body, html {
 
 .dashboard {
   display: flex;
-  height: 100vh; /* Full viewport height */
+  height: 100vh;
 }
 
 .sidebar {
-  width: 250px;
+  width: var(--sidebar-width);
   background-color: var(--primary-color);
   color: var(--background-color);
   padding: 20px;
   min-height: 100vh;
+  transition: width 0.3s ease-in-out;
 }
 
 .sidebar ul {
@@ -34,30 +48,35 @@ body, html {
 }
 
 .sidebar-item, .sidebar ul li {
-  padding: 15px 20px;
+  padding: 10px 15px;
   cursor: pointer;
-  border-bottom: 1px solid #444;
-  transition: background-color 0.3s;
+  border-bottom: 1px solid #444; 
+  transition: background-color 0.3s ease-in-out;
+  font-size: var(--sidebar-font-size); 
 }
 
 .sidebar-item:hover, .sidebar-item:focus, .sidebar ul li:hover, .sidebar ul li:focus {
   background-color: #555;
+  color: white;
 }
 
-.main-content, .campaign-table {
+.main-content {
   flex-grow: 1;
   background-color: var(--background-color);
-  overflow-y: auto; /* Enables scroll if content is taller than screen */
+  overflow-y: auto;
   padding: 20px;
 }
 
-/* Table styles */
-table, .campaign-table table {
+.campaign {
+  animation: slideInFromLeft 0.5s ease-out forwards;
+}
+
+table {
   width: 100%;
   border-collapse: collapse;
 }
 
-th, td, .campaign-table th, .campaign-table td {
+th, td {
   text-align: left;
   padding: 8px;
   border-bottom: 1px solid var(--border-color);
@@ -67,7 +86,6 @@ th {
   background-color: #eee;
 }
 
-/* Enhanced search bar and filters styling */
 .search-and-filters {
   display: flex;
   padding: 20px;
@@ -82,13 +100,40 @@ th {
   border: 1px solid var(--border-color);
   background-color: var(--input-background);
   border-radius: var(--input-border-radius);
-  flex-grow: 1; /* Search bar takes more space */
+  flex-grow: 1;
 }
 
 .search-and-filters select {
-  flex-grow: 0; /* Dropdowns do not expand */
+  flex-grow: 0;
 }
 
 .search-and-filters select:last-child {
   margin-right: 0;
+}
+
+button, .button {
+  background-color: var(--highlight-color);
+  color: white;
+  border: none;
+  padding: 10px 15px;
+  border-radius: var(--input-border-radius);
+  cursor: pointer;
+  transition: background-color 0.2s ease-in-out;
+}
+
+button:hover, .button:hover {
+  background-color: darken(var(--highlight-color), 10%);
+}
+
+@media screen and (max-width: 768px) {
+  .sidebar {
+    width: 100%;
+    height: auto;
+    position: relative;
+    padding: 10px;
+  }
+
+  .main-content {
+    padding: 10px;
+  }
 }
